requirements.txt
boto3==1.27.1
kafka-python==2.0.2